# contracts/openapi/platform-draft.yaml
openapi: 3.1.0
info:
  title: OpsWeave platform API
  version: 0.1.0-draft
  description: >
    Identity-scoped business endpoints. Tenant and permissions are taken from the
    trusted principal, never from query or body. Explicit OIDC mode uses the authorization-code BFF;
    dev mode accepts only its configured loopback credential. Cookie requests require exact Origin
    and X-CSRF-TOKEN on unsafe methods; GET session metadata supplies the masked token.
    Zabbix Host ingest is available only when an explicit source mode is configured.
    All API responses use Cache-Control no-store and an X-OpsWeave-Request-Id UUID.
    The optional client request ID is correlation metadata only, not identity or idempotency.
paths:
...
  /api/v1/entities/{entityId}/relations:
    parameters:
      - {$ref: '#/components/parameters/clientRequestId'}
      - {name: entityId, in: path, required: true, schema: {type: string, format: uuid}}
    get:
      operationId: pageEntityRelations
      description: Read active relation instances touching the path entity at the requested asOf, or at the server clock when omitted. Tenant, entity scope and model definitions are resolved from the trusted principal. The page is live and bounded; it does not create or infer relations.
      security: [{bearerAuth: []}, {browserSession: []}]
      parameters:
        - {name: after, in: query, schema: {type: string, format: uuid}}
        - {name: asOf, in: query, schema: {type: string, format: date-time}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 50, default: 25}}
      responses:
        '200':
          description: Authorized active relation instances
          content:
            application/json:
              schema: {$ref: ../schemas/v1/entity-relation-page.schema.json}
        '400': {description: Invalid entity, cursor, asOf or limit}
        '401': {description: Unauthenticated}
        '403': {description: Missing entity.read or object scope}
        '404': {description: Entity unavailable in the trusted tenant}
        '503': {description: Relation storage unavailable; no empty fallback}
    post:
      operationId: createEntityRelation
      description: Create one relation instance using a published immutable relation model. entity.manage is required on both endpoints; endpoint types and versions are checked before the insert. requestId is the idempotency key, retries return the original receipt, and the request never accepts a tenant override. validFrom cannot be materially in the future and validTo must be after it.
      security: [{bearerAuth: []}, {browserSession: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: ../schemas/v1/entity-relation-command.schema.json}
      responses:
        '200':
          description: Durable relation write receipt; sourceRef is retained server-side and is not returned
          content:
            application/json:
              schema: {$ref: ../schemas/v1/entity-relation-receipt.schema.json}
        '400': {description: Invalid closed JSON, endpoint path mismatch, time or version}
        '401': {description: Unauthenticated}
        '403': {description: Missing entity.manage or endpoint scope}
        '404': {description: Endpoint unavailable in the trusted tenant}
        '409': {description: Request reuse, stale endpoint version, unpublished model or endpoint type mismatch}
        '503': {description: Relation storage unavailable; resolve by requestId before retry}
...
  /api/v2/data-sources:
    parameters:
      - {$ref: '#/components/parameters/clientRequestId'}
    get:
      operationId: pageSourceInstances
      description: Bounded owner-scoped source instance page. The server derives tenant, subject and permissions; no network probe or collection is started.
      security: [{bearerAuth: []}, {browserSession: []}]
      responses:
        '200':
          description: The newest visible source instances
          content:
            application/json:
              schema: {$ref: '../schemas/v2/source-instance-page.schema.json'}
        default: {description: 400 unknown query; 401 unauthenticated; 403 denied; 503 unavailable}
    post:
      operationId: createSourceInstance
      description: Confirm one explicitly selected registered source setup. The request is idempotent by owner and requestId and creates only the initial instance projection; it does not start collection.
      security: [{bearerAuth: []}, {browserSession: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '../schemas/v2/source-instance-create-command.schema.json'}
      responses:
        '200':
          description: Initial instance snapshot
          content:
            application/json:
              schema: {$ref: '../schemas/v2/source-instance-read.schema.json'}
        default: {description: 400 invalid closed JSON; 403 denied; 409 reused request or capacity; 503 unavailable}
...
  /api/v2/data-sources/connections:
    parameters:
      - {$ref: '#/components/parameters/clientRequestId'}
    post:
      operationId: createSourceConnection
      description: Create a source instance and append its first endpoint/credential pin in one explicit transaction. Pins are checked by the trusted server; no secret is returned.
      security: [{bearerAuth: []}, {browserSession: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '../schemas/v2/source-connection-create.schema.json'}
      responses:
        '200':
          description: Connection and instance receipt
          content:
            application/json:
              schema: {$ref: '../schemas/v2/source-connection-receipt.schema.json'}
        default: {description: 400 invalid closed JSON; 403 denied; 409 stale or reused request; 503 unavailable}
  /api/v2/data-sources/{id}/connection:
    parameters:
      - {$ref: '#/components/parameters/clientRequestId'}
      - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
    get:
      operationId: readSourceConnection
      description: Read current connection metadata and availability. Endpoint address is metadata only; this GET never probes it.
      security: [{bearerAuth: []}, {browserSession: []}]
      responses:
        '200':
          description: Connection view without secret material
          content:
            application/json:
              schema: {$ref: '../schemas/v2/source-connection-read.schema.json'}
        default: {description: 400 invalid id; 403 denied; 404 missing; 503 unavailable}
    patch:
      operationId: editSourceConnection
      description: Append a new immutable endpoint/credential pin under an instance edit CAS. The server rechecks both pins and never exposes the secret.
      security: [{bearerAuth: []}, {browserSession: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '../schemas/v2/source-connection-edit.schema.json'}
      responses:
        '200':
          description: Connection update receipt
          content:
            application/json:
              schema: {$ref: '../schemas/v2/source-connection-receipt.schema.json'}
        default: {description: 400 invalid closed JSON; 403 denied; 409 stale pin/version or reused request; 503 unavailable}
  /api/v2/data-sources/{id}/connection/history:
    parameters:
      - {$ref: '#/components/parameters/clientRequestId'}
      - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
    get:
      operationId: pageSourceConnectionHistory
      description: Read-only append-only connection configuration history, bounded to 100 revisions.
      security: [{bearerAuth: []}, {browserSession: []}]
      responses:
        '200':
          description: Connection history
          content:
            application/json:
              schema: {$ref: '../schemas/v2/source-connection-history.schema.json'}
        default: {description: 400 invalid id; 403 denied; 404 missing; 503 unavailable}
  /api/v2/data-sources/{id}/connection/commands/{requestId}:
    parameters:
      - {$ref: '#/components/parameters/clientRequestId'}
      - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
      - {name: requestId, in: path, required: true, schema: {type: string, format: uuid}}
    get:
      operationId: getSourceConnectionCommandReceipt
      description: Read the original connection command receipt after an uncertain response; it does not repeat the write.
      security: [{bearerAuth: []}, {browserSession: []}]
      responses:
        '200':
          description: Original connection receipt
          content:
            application/json:
              schema: {$ref: '../schemas/v2/source-connection-receipt.schema.json'}
        '404': {description: Missing or hidden receipt}
        default: {description: 400 invalid id; 403 denied; 503 unavailable}
...
components:
  parameters:
    clientRequestId:
      in: header
      name: X-OpsWeave-Request-Id
      required: false
      description: Optional single UUID; malformed or duplicate values return 400. Server generates one if absent and echoes a canonical UUID on every API response. Never grants permissions or suppresses duplicate writes.
      schema: {type: string, format: uuid}
      example: 6c8d4385-c206-4b59-8c2e-f9d34f8b3500
...
  securitySchemes:
    historyService:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Dedicated OAuth client_credentials access token profile described in ../history-service-identity.md. Never accepted as a browser or Runtime credential.
    runtimeKey:
      type: apiKey
      in: header
      name: X-OpsWeave-Runtime-Key
      description: Explicit local process attestation only; does not grant user permissions and is never exposed to Web/model.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Explicit loopback dev credential, or short-lived server-only Runtime delegation on allowlisted endpoints. Raw provider bearer JWTs are not accepted.
    browserSession:
      type: apiKey
      in: cookie
      name: __Host-opsweave
      description: OIDC BFF HttpOnly Secure SameSite=Lax host cookie. Explicit loopback protocol tests use opsweave-test-session. Unsafe methods also require X-CSRF-TOKEN and the configured Origin.